+++ src/webviews/extension-side/plotting/plotViewer.node.ts
@@ -5,121 +5,35 @@ import '../../../platform/common/extensions';
 
 import { inject, injectable } from 'inversify';
 import * as path from '../../../platform/vscode-path/path';
-import { Event, EventEmitter, Uri, ViewColumn } from 'vscode';
-
+import { Uri } from 'vscode';
 import { traceError, traceInfo } from '../../../platform/logging';
 import { createDeferred } from '../../../platform/common/utils/async';
-import { PlotViewerMessageListener } from './plotViewerMessageListener';
-import { IExportPlotRequest, IPlotViewer, IPlotViewerMapping, PlotViewerMessages } from './types';
-import {
-    IWebviewPanelProvider,
-    IWorkspaceService,
-    IApplicationShell
-} from '../../../platform/common/application/types';
-import { IConfigurationService, IDisposable, IExtensionContext } from '../../../platform/common/types';
-import { IFileSystem } from '../../../platform/common/platform/types';
+import { IExportPlotRequest } from './types';
+import { IFileSystemNode } from '../../../platform/common/platform/types.node';
 import * as localize from '../../../platform/common/utils/localize';
-import { WebviewPanelHost } from '../../../platform/webviews/webviewPanelHost';
-import { joinPath } from '../../../platform/vscode-path/resources';
 import { noop } from '../../../platform/common/utils/misc';
+import { PlotViewer as PlotViewerBase } from './plotViewer';
+import {
+    IApplicationShell,
+    IWebviewPanelProvider,
+    IWorkspaceService
+} from '../../../platform/common/application/types';
+import { IConfigurationService, IExtensionContext } from '../../../platform/common/types';
 
 @injectable()
-export class PlotViewer extends WebviewPanelHost<IPlotViewerMapping> implements IPlotViewer, IDisposable {
-    private closedEvent: EventEmitter<IPlotViewer> = new EventEmitter<IPlotViewer>();
-    private removedEvent: EventEmitter<number> = new EventEmitter<number>();
-
+export class PlotViewer extends PlotViewerBase {
     constructor(
         @inject(IWebviewPanelProvider) provider: IWebviewPanelProvider,
         @inject(IConfigurationService) configuration: IConfigurationService,
         @inject(IWorkspaceService) workspaceService: IWorkspaceService,
-        @inject(IApplicationShell) private applicationShell: IApplicationShell,
-        @inject(IFileSystem) private fs: IFileSystem,
-        @inject(IExtensionContext) readonly context: IExtensionContext
+        @inject(IApplicationShell) applicationShell: IApplicationShell,
+        @inject(IFileSystemNode) private fsNode: IFileSystemNode,
+        @inject(IExtensionContext) context: IExtensionContext
     ) {
-        const plotDir = joinPath(context.extensionUri, 'out', 'webviews', 'webview-side', 'viewers');
-        super(
-            configuration,
-            provider,
-            workspaceService,
-            (c, v, d) => new PlotViewerMessageListener(c, v, d),
-            plotDir,
-            [joinPath(plotDir, 'plotViewer.js')],
-            localize.DataScience.plotViewerTitle(),
-            ViewColumn.One
-        );
-        // Load the web panel using our current directory as we don't expect to load any other files
-        super.loadWebview(Uri.file(process.cwd())).catch(traceError);
+        super(provider, configuration, workspaceService, applicationShell, fsNode, context);
     }
 
-    public get closed(): Event<IPlotViewer> {
-        return this.closedEvent.event;
-    }
-
-    public get removed(): Event<number> {
-        return this.removedEvent.event;
-    }
-
-    public override async show(): Promise<void> {
-        if (!this.isDisposed) {
-            // Then show our web panel.
-            return super.show(true);
-        }
-    }
-
-    public addPlot = async (imageHtml: string): Promise<void> => {
-        if (!this.isDisposed) {
-            // Make sure we're shown
-            await super.show(false);
-
-            // Send a message with our data
-            this.postMessage(PlotViewerMessages.SendPlot, imageHtml).ignoreErrors();
-        }
-    };
-
-    public override dispose() {
-        super.dispose();
-        if (this.closedEvent) {
-            this.closedEvent.fire(this);
-        }
-    }
-
-    protected get owningResource() {
-        return undefined;
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    protected override onMessage(message: string, payload: any) {
-        switch (message) {
-            case PlotViewerMessages.CopyPlot:
-                this.copyPlot(payload.toString()).ignoreErrors();
-                break;
-
-            case PlotViewerMessages.ExportPlot:
-                this.exportPlot(payload).ignoreErrors();
-                break;
-
-            case PlotViewerMessages.RemovePlot:
-                this.removePlot(payload);
-                break;
-
-            default:
-                break;
-        }
-
-        super.onMessage(message, payload);
-    }
-
-    private removePlot(payload: number) {
-        this.removedEvent.fire(payload);
-    }
-
-    private copyPlot(_svg: string): Promise<void> {
-        // This should be handled actually in the web view. Leaving
-        // this here for now in case need node to handle it.
-        return Promise.resolve();
-    }
-
-    private async exportPlot(payload: IExportPlotRequest): Promise<void> {
+    protected override async exportPlot(payload: IExportPlotRequest): Promise<void> {
         traceInfo('exporting plot...');
         const filtersObject: Record<string, string[]> = {};
         filtersObject[localize.DataScience.pdfFilter()] = ['pdf'];
@@ -136,7 +50,7 @@ export class PlotViewer extends WebviewPanelHost<IPlotViewerMapping> implements 
                 const ext = path.extname(file.fsPath);
                 switch (ext.toLowerCase()) {
                     case '.pdf':
-                        await saveSvgToPdf(payload.svg, this.fs, file);
+                        await saveSvgToPdf(payload.svg, this.fsNode, file);
                         break;
 
                     case '.png':
@@ -158,7 +72,7 @@ export class PlotViewer extends WebviewPanelHost<IPlotViewerMapping> implements 
     }
 }
 
-export async function saveSvgToPdf(svg: string, fs: IFileSystem, file: Uri) {
+export async function saveSvgToPdf(svg: string, fs: IFileSystemNode, file: Uri) {
     traceInfo('Attempting pdf write...');
     // Import here since pdfkit is so huge.
     // eslint-disable-next-line @typescript-eslint/no-require-imports
